requirements.txt
requests
beautifulsoup4
lxml
discord.py
discord-webhook